refactor: Перенести контроллер авторизации в контроллер users
