+++ controllers/users.js
@@ -1,6 +1,10 @@
+const jwt = require('jsonwebtoken');
+const bcrypt = require('bcryptjs');
 const User = require('../models/user');
-const { NotFoundError } = require('../errors/NotFoundError');
 const { BadRequestError } = require('../errors/BadRequestError');
+const { ConflictError } = require('../errors/ConflictError');
+const { UnauthorizedError } = require('../errors/UnauthorizedError');
+const { NotFoundError } = require('../errors/NotFoundError');
 
 function getCurrentUser(req, res, next) {
   const userId = req.user._id;
@@ -46,6 +50,47 @@ function updateUserInfo(req, res, next) {
     });
 }
 
+function register(req, res, next) {
+  const { email, password } = req.body;
+
+  if (!email || !password) throw new BadRequestError('Email или пароль не могут быть пустыми');
+
+  bcrypt.hash(password, 10)
+    .then((hash) => {
+      User.create({
+        email, password: hash,
+      })
+        .then(() => res.status(200).send({ email }))
+        .catch((err) => {
+          if (err.code === 11000) {
+            return next(new ConflictError('Пользователь с таким email уже существует'));
+          }
+          if (err.name === 'ValidationError') {
+            return next(new BadRequestError('Введены некорретные данные'));
+          }
+          next(err);
+        });
+    }).catch(next);
+}
+
+function login(req, res, next) {
+  const { email, password } = req.body;
+
+  return User.findUserByCredentials(email, password)
+    .then((user) => {
+      const token = jwt.sign(
+        { _id: user._id },
+        'super-strong-secret',
+        { expiresIn: '7d' },
+      );
+
+      res.send({ token });
+    })
+    .catch((err) => {
+      next(new UnauthorizedError(err.message));
+    });
+}
+
 module.exports = {
-  getCurrentUser, updateUserInfo,
+  getCurrentUser, updateUserInfo, register, login,
 };
